# web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "F1 Data Platform API",
    "summary": "Resultados, vueltas, neumáticos, telemetría y estadísticas de la F1 (1950-hoy).",
    "description": "API de solo lectura sobre el modelo dimensional del proyecto. Fuentes: F1DB, FastF1, OpenF1, formula1db.com y Ergast (solo validación), conciliadas con documentos de la FIA. Los tiempos van en milisegundos y los identificadores son los de F1DB (p. ej. `lewis-hamilton`, `ferrari`).\n\nCódigo MIT. Los datos conservan la licencia de su fuente: F1DB (CC BY 4.0), Ergast (CC BY-NC-SA 3.0), OpenF1 (CC BY-NC-SA 4.0: lo que lleva `source = openf1`), el cronometraje de la F1 obtenido con FastF1 (© Formula One, uso académico no comercial) y formula1db.com (con permiso de su autor); ver [licencias de los datos](https://github.com/AntonioSchez32/f1-data-platform#licencia-y-atribuciones). Cita «F1 Data Platform» y la fuente original. Proyecto académico no oficial, sin relación con la Fórmula 1 ni con la FIA.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/seasons/{season}/standings/drivers": {
      "get": {
        "tags": [
          "Temporadas"
        ],
        "summary": "Clasificación de pilotos (tras la última carrera o la ronda indicada)",
        "operationId": "driver_standings_seasons__season__standings_drivers_get",
        "parameters": [
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          },
          {
            "name": "round",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tras esta ronda",
              "title": "Round"
            },
            "description": "Tras esta ronda"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverStanding"
                  },
                  "title": "Response Driver Standings Seasons  Season  Standings Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/drivers": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Buscar pilotos",
        "operationId": "list_drivers_drivers_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "description": "Parte del nombre",
              "title": "Search"
            },
            "description": "Parte del nombre"
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Solo pilotos de esta temporada",
              "title": "Season"
            },
            "description": "Solo pilotos de esta temporada"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverSummary"
                  },
                  "title": "Response List Drivers Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Ficha y estadísticas de carrera",
        "operationId": "get_driver_drivers__driver_id__get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/seasons": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Temporada a temporada",
        "description": "Una fila por temporada con alguna inscripción (decisión 36): la posición es nula si no\nfigura en la clasificación final, y `race_starts` cuenta solo las carreras que corrió.",
        "operationId": "driver_seasons_drivers__driver_id__seasons_get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverSeason"
                  },
                  "title": "Response Driver Seasons Drivers  Driver Id  Seasons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/results": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Resultados carrera a carrera",
        "operationId": "driver_results_drivers__driver_id__results_get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverRaceResult"
                  },
                  "title": "Response Driver Results Drivers  Driver Id  Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/teammates": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Comparativa con cada compañero de equipo (carrera, clasificación y puntos)",
        "description": "Desglose por pareja. El duelo en carrera solo compara las carreras en que acaban los dos\n(`races_both_classified`, decisión 46). Los puntos son los de las carreras con ese compañero:\nno se suman entre parejas (para el total de la temporada, `/teammates/seasons`).",
        "operationId": "teammates_drivers__driver_id__teammates_get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeammateComparison"
                  },
                  "title": "Response Teammates Drivers  Driver Id  Teammates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/teammates/seasons": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Resumen de cada temporada frente a los compañeros",
        "description": "Puntos carrera a carrera frente al mejor compañero de cada carrera (decisión 45) y duelos\nsumados por pareja (decisión 46). Solo las temporadas en que tuvo algún compañero.",
        "operationId": "teammate_seasons_drivers__driver_id__teammates_seasons_get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeammateSeason"
                  },
                  "title": "Response Teammate Seasons Drivers  Driver Id  Teammates Seasons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/teammates/races": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Carrera a carrera frente al mejor compañero (para los acumulados)",
        "description": "Una fila por carrera con compañero. En los años 50 un piloto podía correr con dos equipos\nen la misma carrera: se suman, como en el resumen de la temporada, y `best_teammates` lleva el\nmejor compañero de cada equipo.",
        "operationId": "teammate_races_drivers__driver_id__teammates_races_get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeammateRace"
                  },
                  "title": "Response Teammate Races Drivers  Driver Id  Teammates Races Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/records/drivers": {
      "get": {
        "tags": [
          "Récords"
        ],
        "summary": "Ranking histórico de pilotos",
        "operationId": "driver_records_records_drivers_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "championships",
                "wins",
                "podiums",
                "pole_positions",
                "fastest_laps",
                "race_starts",
                "points",
                "grand_slams"
              ],
              "type": "string",
              "default": "wins",
              "title": "Metric"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordEntry"
                  },
                  "title": "Response Driver Records Records Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/rankings/drivers": {
      "get": {
        "tags": [
          "Récords"
        ],
        "summary": "Pilotos más laureados",
        "operationId": "driver_rankings_rankings_drivers_get",
        "parameters": [
          {
            "name": "season_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Desde esta temporada",
              "title": "Season From"
            },
            "description": "Desde esta temporada"
          },
          {
            "name": "season_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hasta esta temporada",
              "title": "Season To"
            },
            "description": "Hasta esta temporada"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "wins",
                "championships",
                "podiums",
                "pole_positions",
                "fastest_laps",
                "points",
                "entries"
              ],
              "type": "string",
              "default": "wins",
              "title": "Order By"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverRanking"
                  },
                  "title": "Response Driver Rankings Rankings Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriverRaceResult": {
        "properties": {
          "race_id": {
            "type": "integer",
            "title": "Race Id"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "grand_prix_name": {
            "type": "string",
            "title": "Grand Prix Name"
          },
          "session": {
            "type": "string",
            "title": "Session"
          },
          "constructor_id": {
            "type": "string",
            "title": "Constructor Id"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "position_text": {
            "type": "string",
            "title": "Position Text"
          },
          "grid_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid Position"
          },
          "points": {
            "type": "number",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "race_id",
          "season",
          "round",
          "grand_prix_name",
          "session",
          "constructor_id",
          "position",
          "position_text",
          "grid_position",
          "points"
        ],
        "title": "DriverRaceResult"
      },
...
      "DriverSeason": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "constructor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constructor Id"
          },
          "constructor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constructor Name"
          },
          "championship_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Championship Position"
          },
          "championship_position_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Championship Position Text"
          },
          "points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "is_champion": {
            "type": "boolean",
            "title": "Is Champion"
          },
          "races": {
            "type": "integer",
            "title": "Races"
          },
          "race_entries": {
            "type": "integer",
            "title": "Race Entries"
          },
          "race_starts": {
            "type": "integer",
            "title": "Race Starts"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "podiums": {
            "type": "integer",
            "title": "Podiums"
          },
          "pole_positions": {
            "type": "integer",
            "title": "Pole Positions"
          },
          "fastest_laps": {
            "type": "integer",
            "title": "Fastest Laps"
          },
          "best_result": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Result"
          }
        },
        "type": "object",
        "required": [
          "season",
          "constructor_id",
          "constructor_name",
          "championship_position",
          "championship_position_text",
          "points",
          "is_champion",
          "races",
          "race_entries",
          "race_starts",
          "wins",
          "podiums",
          "pole_positions",
          "fastest_laps",
          "best_result"
        ],
        "title": "DriverSeason"
      },
      "DriverStanding": {
        "properties": {
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "position_text": {
            "type": "string",
            "title": "Position Text"
          },
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nationality_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality Alpha2"
          },
          "constructors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constructors"
          },
          "points": {
            "type": "number",
            "title": "Points"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "is_champion": {
            "type": "boolean",
            "title": "Is Champion"
          }
        },
        "type": "object",
        "required": [
          "position",
          "position_text",
          "driver_id",
          "name",
          "nationality_alpha2",
          "constructors",
          "points",
          "wins",
          "is_champion"
        ],
        "title": "DriverStanding"
      },
      "DriverSummary": {
        "properties": {
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nationality_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality Alpha2"
          },
          "first_season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Season"
          },
          "last_season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Season"
          },
          "race_starts": {
            "type": "integer",
            "title": "Race Starts"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "podiums": {
            "type": "integer",
            "title": "Podiums"
          },
          "championships": {
            "type": "integer",
            "title": "Championships"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "name",
          "nationality_alpha2",
          "first_season",
          "last_season",
          "race_starts",
          "wins",
          "podiums",
          "championships"
        ],
        "title": "DriverSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecordEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Alpha2"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "id",
          "name",
          "country_alpha2",
          "value"
        ],
        "title": "RecordEntry"
      },
...
    }
  }
}